# server/openapi.yaml
openapi: 3.0.1
info:
  title: Accountable
  version: 0.5.0
  description: This is the Accountable Server. It doesn't do much, because the Accountable client does most of the hard work. You can read the source code on [GitHub](https://github.com/AverageHelper/accountable-vue).
  license:
    name: GNU General Public License v3.0
    url: https://github.com/AverageHelper/accountable-vue/blob/main/LICENSE
...
servers:
  - url: https://api.accountable.average.name
...
paths:
...
  "/v0/join":
    post:
      tags:
        - auth
      summary: Create a new account
      operationId: create-account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - account
                - password
              type: object
              properties:
                account:
                  type: string
                  description: An identifier (or username) chosen for the user's account.
                    The user may change this later.
                password:
                  type: string
                  description: A passphrase to access the user's account. The user may change this later.
      responses:
        200:
          description: Account has been added to the database with the given credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
                  access_token:
                    type: string
                    description: Temporary bearer token used to authenticate future requests
                  uid:
                    type: string
                    description: The user's persistent UID, different from the given account ID
                  totalSpace:
                    type: number
                    example: 1000000
                    description: "The number of total bytes the server has allocated for the user."
                  usedSpace:
                    type: number
                    example: 85000
                    description: "The number of bytes the user has stored."
        400:
          description: Missing one or more required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Improper parameter types"
        409:
          description: An account with the given ID already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "account-conflict"
                  message:
                    type: string
                    enum:
                      - "An account with that ID already exists"
  "/v0/login":
    post:
      tags:
        - auth
      summary: Get an auth token for an existing account
      operationId: login-account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - account
                - password
              type: object
              properties:
                account:
                  type: string
                  description: An identifier (or username) chosen for the user's account.
                password:
                  type: string
                  description: A passphrase to access the user's account.
      responses:
        200:
          description: Successfully logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
                  access_token:
                    type: string
                    description: Temporary bearer token used to authenticate future requests
                  uid:
                    type: string
                    description: The user's persistent UID, different from the given account ID
                  totalSpace:
                    type: number
                    example: 1000000
                    description: "The number of total bytes the server has allocated for the user."
                  usedSpace:
                    type: number
                    example: 85000
                    description: "The number of bytes the user has stored."
        400:
          description: Missing one or more required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Improper parameter types"
        403:
          description: The given passphrase does not match, or the given account does
            not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "wrong-credentials"
                  message:
                    type: string
                    enum:
                      - "Incorrect account ID or passphrase"
  "/v0/session":
    get:
      tags:
        - auth
      summary: Get a new auth token for an existing session
      operationId: renew-session
      security:
        - jwt: []
      responses:
        200:
          description: Successfully renewed session
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
                  account:
                    type: string
                    description: The ID of the user's account.
                  access_token:
                    type: string
                    description: Temporary bearer token used to authenticate future requests
                  uid:
                    type: string
                    description: The user's persistent UID, different from the given account ID
                  totalSpace:
                    type: number
                    example: 1000000
                    description: "The number of total bytes the server has allocated for the user."
                  usedSpace:
                    type: number
                    example: 85000
                    description: "The number of bytes the user has stored."
        403:
          description: The caller did not send valid session cookies.
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "expired-token"
                      - "missing-token"
                      - "not-owner"
  "/v0/logout":
    post:
      tags:
        - auth
      summary: Invalidate the request's bearer token
      operationId: log-out
      security:
        - jwt: []
      responses:
        200:
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
  "/v0/leave":
    post:
      tags:
        - auth
      summary: Destroy the user's data and account credentials
      operationId: leave-forever
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - account
                - password
              type: object
              properties:
                account:
                  type: string
                  description: The identifier for the user's account.
                password:
                  type: string
                  description: The passphrase to access the account.
      responses:
        200:
          description: Operation succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
        400:
          description: Missing one or more required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Improper parameter types"
        403:
          description: The given passphrase does not match, or the given account does
            not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "wrong-credentials"
                  message:
                    type: string
                    enum:
                      - "Incorrect account ID or passphrase"
  "/v0/updatepassword":
    post:
      tags:
        - auth
      summary: Change the user's password
      operationId: change-password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - account
                - password
                - newpassword
              type: object
              properties:
                account:
                  type: string
                  description: The identifier for the user's account.
                password:
                  type: string
                  description: The current passphrase to access the account.
                newpassword:
                  type: string
                  description: The new passphrase to access the account from now on.
      responses:
        200:
          description: Operation succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
        400:
          description: Missing one or more required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Improper parameter types"
        403:
          description: The given passphrase does not match, or the given account does
            not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "wrong-credentials"
                  message:
                    type: string
                    enum:
                      - "Incorrect account ID or passphrase"
  "/v0/updateaccountid":
    post:
      tags:
        - auth
      summary: Change the user's account ID
      operationId: change-account-id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - account
                - newaccount
                - password
              type: object
              properties:
                account:
                  type: string
                  description: The identifier for the user's account.
                newaccount:
                  type: string
                  description: The new identifier to access the account from now on.
                password:
                  type: string
                  description: The current passphrase to access the account.
      responses:
        200:
          description: Operation succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Success!"
        400:
          description: Missing one or more required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Improper parameter types"
        403:
          description: The given passphrase does not match, or the given account does
            not exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - "wrong-credentials"
                  message:
                    type: string
                    enum:
                      - "Incorrect account ID or passphrase"
...
components:
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT